Split inference to init and predict
