+++ .gitignore
@@ -161,4 +161,4 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 README.md.backup
-src/model_weights
+src/model/model_weights
