+++ src/model/config.json
@@ -1,7 +1,7 @@
 {
-    "OCR_MODEL_PATH": "model_weights/OCR-0.0374.ckpt",
-    "BEAM_MODEL_PATH": "model_weights/beamsearch.arpa",
-    "SEGM_MODEL_PATH": "model_weights/SEGM-2160x3130.pth",
+    "OCR_MODEL_PATH": "model/model_weights/OCR-0.0374.ckpt",
+    "BEAM_MODEL_PATH": "model/model_weights/beamsearch.arpa",
+    "SEGM_MODEL_PATH": "model/model_weights/SEGM-2160x3130.pth",
     "ocr_config":{
         "alphabet": "@ !\"%'()+,-./0123456789:;=?AEFIMNOSTW[]abcdefghiklmnopqrstuvwxyАБВГДЕЖЗИКЛМНОПРСТУХЦЧШЭЮЯабвгдежзийклмнопрстуфхцчшщъыьэюяё№",
         "labels_for_bs": "_@|!\"%'()+,-./0123456789:;=?AEFIMNOSTW[]abcdefghiklmnopqrstuvwxyАБВГДЕЖЗИКЛМНОПРСТУХЦЧШЭЮЯабвгдежзийклмнопрстуфхцчшщъыьэюяё№",
