+++ src/run_model.py
@@ -8,10 +8,11 @@ import cv2
 import numpy as np
 import torch
 from matplotlib import pyplot as plt
+from matplotlib.image import AxesImage as plt_image
 
-import helper_functions as help_fn
-import recognition as rec
-import segmentation as segm
+import model.helper_functions as help_fn
+import model.recognition as rec
+import model.segmentation as segm
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
 warnings.filterwarnings("ignore")
@@ -19,7 +20,7 @@ logger = logging.getLogger("detectron2")
 logger.setLevel(logging.CRITICAL)
 
 
-with open("config.json") as f:
+with open("model/config.json") as f:
     CONFIG_JSON = json.load(f)
 
 CONFIG_JSON["ocr_config"]["device"] = (
@@ -55,16 +56,25 @@ class PiepleinePredictor:
         return output
 
 
-def htr_predict(img_name: str) -> None:
-    pipeline_predictor = PiepleinePredictor(
+def init_predictor() -> PiepleinePredictor:
+    return PiepleinePredictor(
         segm_model_path=CONFIG_JSON["SEGM_MODEL_PATH"],
         ocr_model_path=CONFIG_JSON["OCR_MODEL_PATH"],
         beam_model_path=CONFIG_JSON["BEAM_MODEL_PATH"],
         ocr_config=CONFIG_JSON["ocr_config"],
     )
-    image = cv2.imread(img_name)
+
+
+def htr_predict(
+    img_name: str | np.ndarray, pipeline_predictor: PiepleinePredictor
+) -> plt_image:
+    if type(img_name) == str:
+        image = cv2.imread(img_name)
+    else:
+        image = img_name
     output = pipeline_predictor(image)
-    vis = help_fn.get_image_visualization(image, output, "font.ttf")
+    vis = help_fn.get_image_visualization(image, output, "model/font.ttf")
     plt.figure(figsize=(20, 20))
-    plt.imshow(vis)
-    plt.savefig("pred-" + img_name.split("/")[-1])
+    pred_img = plt.imshow(vis)
+    return pred_img
+    # plt.savefig("pred-" + img_name.split("/")[-1])
